+++ src/components/index.ts
@@ -1,2 +1,4 @@
 export * from './chat-watcher/chat-watcher';
+export * from './element-hider/';
 export * from './settings/';
+