+++ src/components/settings/index.tsx
@@ -1,3 +1,4 @@
+import { getElement } from '@/utils/getElement';
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { MAIN_NAVBAR_TOKEN, SETTINGS_MODAL_ID } from './settings-constants';
@@ -15,27 +16,8 @@ const renderSettingsModal = () => {
   );
 };
 
-const waitForContainer = () => {
-  return new Promise((resolve) => {
-    const container = document.body.querySelector(`.${MAIN_NAVBAR_TOKEN}`);
-    if (container) {
-      resolve(container);
-    } else {
-      const observer = new MutationObserver(() => {
-        const observerContainer = document.body.querySelector(`.${MAIN_NAVBAR_TOKEN}`);
-        if (observerContainer) {
-          observer.disconnect();
-          resolve(observerContainer);
-        }
-      });
-      observer.observe(document.body, { childList: true, subtree: true });
-    }
-  });
-};
-
-waitForContainer().then((container) => {
-  const c = container as Element;
-  c.insertBefore(settingsModalElement, c.children[c.childElementCount - 1]);
+getElement(document.body, `.${MAIN_NAVBAR_TOKEN}`).then((container) => {
+  container.insertBefore(settingsModalElement, container.children[container.childElementCount - 1]);
   renderSettingsModal();
 });