+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added 'display videos' which gives basic info about linked video (from streamable/youtube) directly in the chat
+- Added option to hide 'top gifters' from the chat
+- Added generic function for grabbing elements from the page
 
 ### Changed