+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-07-29
+
+### Added
+
+- Added 'display videos' which gives basic info about linked video (from streamable/youtube) directly in the chat
+- Added option to hide 'top gifters' from the chat
+- Added option to hide 'quick emote holder' from the chat
+- Added generic function for grabbing elements from the page
+
+### Changed
+
+- Changed settings-manager to be more bullet proof
+- Changed file structure and naming for a better DX
+
 ## [1.0.0] - 2023-07-20
 
 ### Added